Move from domain.exceptions to exceptions
